feat: allow fractional desired task counts

+++ examples/work_unit_feed.rs
@@ -252,7 +252,9 @@ fn remote_scan_desired_task_count_handler(
         .feed
         .inner()?
         .task_count;
-    Some(Ok(DesiredTaskCountEventResponse::desired(task_count)))
+    Some(Ok(DesiredTaskCountEventResponse::desired(
+        task_count as f64,
+    )))
 }
 
 fn remote_scan_scale_up_leaf_node_handler(
